// src/common/enums/entity.enum.ts
export enum EntityName {
  User = "user",
  Profile = "profile",
  Otp = "otp",
  Category = "category",
  News = "news",
  NewsCategory = "news_category",
  NewsLikes = "news_likes",
  NewsComments = "news_comments",
  NewsBookmark = "news_bookmarks",
  NewsTag = "tags",
  NewsTags = "news_tag",
  Image = "image",
}